pyproject.toml
[tool.poetry]
name = "Python test assessment by DevelopsToday"
version = "0.1.0"
description = ""
authors = ["Kovalev Evgeniy"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
SQLAlchemy = "^1.4.2"
psycopg2-binary = "^2.8.6"
django-cors-headers = "^3.7.0"
djangorestframework = "^3.12.2"
django-extensions = "^3.1.1"
celery = "^5.0.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
flake8-isort = "^4.0.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
